requirement.txt
langchain
langchain_openai
langchain_community
langchain_core
langchain-pinecone
pinecone
streamlit
python-dotenv
PyPDF2
fastapi

-e.